+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log of DVH Analytics
 
+v0.6.2 (2019.12.08)
+--------------------
+ - [Machine Learning] More visual options available in settings
+ - [Plot Visuals] Hover information pop-ups display with data only. Trends, stats, etc. no longer clutter
+ - [Stats Data] Spreadsheet class created, used to display Stats data with ability to edit. 
+ This means you can add custom numerical data for trending, regressions, machine learning.
+ - [Misc] All data, multi-variable regression, and machine learning windows close when main window is closed
+
 v0.6.1 (2019.7.10)
 --------------------
  - [Control Chart] Adjusted Control Charts automatically populated from Regressions tab after running model